Fixd mypy and pylint, increased tolerance threshold
