+++ allennlp/common/testing/__init__.py
@@ -1,1 +1,2 @@
 from allennlp.common.testing.test_case import AllenNlpTestCase
+from allennlp.common.testing.model_test_case import ModelTestCase
